refactor tokenizer
